+++ json_serializable/lib/src/type_helpers/json_helper.dart
@@ -18,6 +18,7 @@ import 'generic_factory_helper.dart';
 
 const _helperLambdaParam = 'value';
 
+/// Supports types that have `fromJson` constructors and/or `toJson` functions.
 class JsonHelper extends TypeHelper<TypeHelperContextWithConfig> {
   const JsonHelper();
 
@@ -131,7 +132,7 @@ class JsonHelper extends TypeHelper<TypeHelperContextWithConfig> {
     // https://github.com/google/json_serializable.dart/issues/19
     output = '${targetType.element.name}.fromJson($output)';
 
-    return DefaultContainer(targetType.isNullableType, expression, output);
+    return DefaultContainer(expression, output);
   }
 }