+++ json_serializable/test/default_value/default_value.g.dart
@@ -37,6 +37,13 @@ DefaultValue _$DefaultValueFromJson(Map<String, dynamic> json) => DefaultValue(
       constClass: json['constClass'] == null
           ? const ConstClass('value')
           : ConstClass.fromJson(json['constClass'] as Map<String, dynamic>),
+      valueFromConverter: json['valueFromConverter'] == null
+          ? const ConstClass('value')
+          : const ConstClassConverter()
+              .fromJson(json['valueFromConverter'] as String),
+      valueFromFunction: json['valueFromFunction'] == null
+          ? const ConstClass('value')
+          : constClassFromJson(json['valueFromFunction'] as String),
     );
 
 Map<String, dynamic> _$DefaultValueToJson(DefaultValue instance) =>
@@ -54,6 +61,9 @@ Map<String, dynamic> _$DefaultValueToJson(DefaultValue instance) =>
       'fieldMapListString': instance.fieldMapListString,
       'fieldEnum': _$GreekEnumMap[instance.fieldEnum],
       'constClass': instance.constClass,
+      'valueFromConverter':
+          const ConstClassConverter().toJson(instance.valueFromConverter),
+      'valueFromFunction': constClassToJson(instance.valueFromFunction),
     };
 
 K _$enumDecode<K, V>(